Add files via upload
The utility of certain operators such as +,<,>,=,/,* and so on is predefined. Programmers can use them directly on built-in data types to write programs.Operator overloading extends the meaning of existing operators so that in addition to the basic data types, they can also be applied to user_defined data types.